requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic==2.12.5
requests==2.32.5

